TASK: Implement NodeMigrationGeneratorService
